+++ RELEASE_NOTES.md
@@ -2,7 +2,13 @@
 
 ## Changes
 
+- **Backglass Manager**: Added upload options via dialog and drag-and-drop:
+
+  <img src="https://raw.githubusercontent.com/syd711/vpin-studio/main/documentation/tables/backglass-manager-dnd.png" width="700" />
+
+
 ## Bugfixes
 
 - **Auto-Shutdown**: Fixed resetting of auto-shutdown on key events. This reset got lost when changing to the new controller input API. (And I had such a great start on "Attack From Mars" when the shutdown warning popped up). 
 - **Highscores**: Fixed highscore resolving for **White Water** by adding "RIVER MASTER" to the list of highscore titles.
+- **Highscores**: Fixed ROM resolving and highscore parsing for **American Most Haunted**.